refactor: Relocate Migration and Notification components to Services namespace
Moves database migration files and all notification-related components to the `server.core/Services` directory and updates their namespaces accordingly. This refactoring improves code organization by consolidating service-oriented code under a dedicated top-level namespace.
