+++ server.core/Services/Notification/Email/EmailAddressListAttribute.cs
@@ -1,6 +1,6 @@
 using System.ComponentModel.DataAnnotations;
 
-namespace Server.Core.Notification;
+namespace server.core.Services.Notification.Email;
 
 [AttributeUsage(AttributeTargets.Property | AttributeTargets.Field | AttributeTargets.Parameter, AllowMultiple = false)]
 public sealed class EmailAddressListAttribute : ValidationAttribute
