+++ server.core/Services/Notification/Email/EmailTransportSecurity.cs
@@ -1,6 +1,6 @@
 using MailKit.Security;
 
-namespace Server.Core.Notification;
+namespace server.core.Services.Notification.Email;
 
 internal static class EmailTransportSecurity
 {
