+++ server.core/Services/Notification/Email/EmailValidation.cs
@@ -1,6 +1,6 @@
 using System.ComponentModel.DataAnnotations;
 
-namespace Server.Core.Notification;
+namespace server.core.Services.Notification.Email;
 
 internal static class EmailValidation
 {
